homework week two
*note to Andrew: by chance, I looked up Boston's site, and it turns out, they've updated since I last checked and basically created the type of site I was pitching!! Therefore, I scrambled to make this new idea. I hope that's okay with you.

Lines 2-4 specify font style, size, and text alignment for the body of any page.
Lines 9-14 specify how the link to login will be: solid border around, and block element taking up 40% of the page. 
Lines 19-23 say link to login should turn the border and box green when hovering over it.
Lines 27-29 specify that links should be black, bolded text, with no default underline.
Lines 34-38 specify how, when someone hovers over a link, a green box should appear, taking up 40% the width of the page.

Log in/About page:
Line 43 says that text in login box should be size 14.
Lines 48-52 say the login box should be centered on the page, taking up 30% of the width, the text should be left aligned, and the border should be dark green.
Line 55 says padding for text in the box should be 10px.
Line 59 says center the heading in the box.
Lines 67-70 make the image of the logo centered in the box, and take up 20% of the page width.
Lines 75-76 center the information on the log in page, and make it take up 80% of the page width.
Line 82 puts padding of 30px for that info.

#Sign up page
Lines 87-92 center the sign up box, telling it to take up 40% of the page width, with centered text and a black border around it.